+++ src/pages/folderListPage/FolderListPage.jsx
@@ -1,9 +1,9 @@
-import { getFolders } from "../api/api";
+import { getFolders } from "../../api/api";
 import { useEffect, useState } from "react";
-import searchBarStyles from "../components/SearchBar.module.css";
-import ListPage from "../components/ListPage";
-import LinkItem from "../components/LinkItem";
-import searchIcon from "../assets/Search.png";
+import searchBarStyles from "../../components/searchBar/SearchBar.module.css";
+import ListPage from "../../components/listPage/ListPage";
+import LinkItem from "../../components/linkItem/LinkItem";
+import searchIcon from "../../assets/Search.png";
 import styles from "./FolderListPage.module.css";
 
 function FolderListPage() {
